Deny code after STAGE, read type, nuke syntax test
Once the user has used a STAGE, everything after it must be enclosed in
a STAGE as well.
For here onward we read the full type of our stream tokens.
Finally, the syntax tests have been moved to inside
`src/spar_stream.rs`, so we don't need them in `tests/syntax` anymore.